handle signout 6/3/2021

+++ src/actions/auth.actions.js
@@ -1,19 +1,19 @@
-import axios from "../helper/axios";
 import { authConstants } from "./constants";
+import axios from "../helper/axios";
 
 export const login = (user) => {
+  console.log(user);
+
   return async (dispatch) => {
     dispatch({ type: authConstants.LOGIN_REQUEST });
-    const res = await axios.post("/admin/signin", {
+    const res = await axios.post(`/admin/signin`, {
       ...user,
     });
+
     if (res.status === 200) {
       const { token, user } = res.data;
-
       localStorage.setItem("token", token);
-
       localStorage.setItem("user", JSON.stringify(user));
-
       dispatch({
         type: authConstants.LOGIN_SUCCESS,
         payload: {
@@ -35,10 +35,8 @@ export const login = (user) => {
 export const isUserLoggedIn = () => {
   return async (dispatch) => {
     const token = localStorage.getItem("token");
-
     if (token) {
       const user = JSON.parse(localStorage.getItem("user"));
-
       dispatch({
         type: authConstants.LOGIN_SUCCESS,
         payload: {
@@ -49,7 +47,7 @@ export const isUserLoggedIn = () => {
     } else {
       dispatch({
         type: authConstants.LOGIN_FAILURE,
-        payload: { error: "failed to login" },
+        payload: { error: "Failed to login" },
       });
     }
   };
@@ -57,9 +55,17 @@ export const isUserLoggedIn = () => {
 
 export const signout = () => {
   return async (dispatch) => {
-    localStorage.clear();
-    dispatch({
-      type: authConstants.LOGOUT_REQUEST,
-    });
+    dispatch({ type: authConstants.LOGOUT_REQUEST });
+    const res = await axios.post(`/admin/signout`);
+
+    if (res.status === 200) {
+      localStorage.clear();
+      dispatch({ type: authConstants.LOGOUT_SUCCESS });
+    } else {
+      dispatch({
+        type: authConstants.LOGOUT_FAILURE,
+        payload: { error: res.data.error },
+      });
+    }
   };
 };
